유저 프로필 페이지 수정, 비밀번호 수정 기능 추가

+++ controllers/videoController.js
@@ -60,7 +60,10 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -71,7 +74,7 @@ export const videoDetail = async (req, res) => {
     },
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", {
       pageTitle: video.title,
       video,
@@ -89,10 +92,14 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", {
-      pageTitle: `${video.title} 수정하기`,
-      video,
-    });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", {
+        pageTitle: `${video.title} 수정하기`,
+        video,
+      });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -127,9 +134,16 @@ export const deleteVideo = async (req, res) => {
     },
   } = req;
   try {
-    await Video.findOneAndRemove({
-      _id: id,
-    });
-  } catch (error) {}
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({
+        _id: id
+      });
+    }
+  } catch (error) {
+    console.log(error);
+  }
   res.redirect(routes.home);
 };
